requirements.txt
mcp>=1.8.0
lxml>=4.9.2
python-pptx>=0.6.21
reportlab>=4.0.0
svglib>=1.5.1